Moved tasks render from module "task" to "board"
